docs: 지출 기록 상세보기 API 문서

+++ Module-API/src/main/java/depromeet/api/domain/feed/controller/FeedController.java
@@ -50,7 +50,7 @@ public class FeedController {
             })
     @GetMapping("/challenge/{challengeRoomId}")
     public Response<GetChallengeRoomResponse> getChallengeRoom(
-            @PathVariable("challengeRoomId") Integer challengeRoomId) {
+            @PathVariable("challengeRoomId") Long challengeRoomId) {
 
         return ResponseService.getDataResponse(getChallengeRoomUseCase.execute());
     }
