+++ Module-API/src/main/java/depromeet/api/domain/feed/dto/ParticipatedChallenge.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class ParticipatedChallenge {
 
     @Schema(example = "13")
-    private Integer challengeRoomId;
+    private Long challengeRoomId;
 
     @Schema(example = "배달 10만원 이하로 쓰기")
     private String title;
